Usable Capacity for Snowball Edge Correction
There is ambiguity in the correct Usable Capacity of the Snowball Edge. This page says that the capacity is 83 TB but based on the official Snowball Edge FAQs and other articles, it is 80 TB: 

https://docs.aws.amazon.com/snowball/latest/developer-guide/specifications.html#specs-storage-optimized


Snowball FAQ: 

Q: What is the difference between the Snowball Edge Storage Optimized and Snowball Edge Compute Optimized options?

Snowball Edge Storage Optimized is the optimal choice if you need to securely and quickly transfer dozens of terabytes to petabytes of data to AWS. It is also a good fit for running general purpose analysis such as IoT data aggregation and transformation. It provides up to 80 TB of usable HDD storage.

Reference: 
https://aws.amazon.com/snowball/faqs/